HACK: tolerate files with more than one sourceMappingURL line
Javascript files may have "sourceMappingURL" as part of their source (eg.
in a string). This is explicitly not supported by the now-deprecated
source-map-resolve module:

"So why is source-map-resolve deprecated? Well, source-map-resolve tries
to work with source maps completely in general – support every feature
and be completely language agnostic. That is a cool goal, but has a
couple of flaws:
- Finding source map comments. This is done via regex to support any
language, but in reality you need to know the language to know what
a comment is, as opposed to something that looks like a comment inside
a string."
https://github.com/lydell/source-map-resolve

source-map-resolve will take the first match, but if more than one exists,
usually the one last will be correct.

This is a workaround for that case.

Signed-off-by: Silvio Moioli <silvio@moioli.net>